requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
numpy==1.26.4
scikit-learn==1.3.0

# ML Models - Python 3.11.9 Compatible
torch==2.4.1
transformers==4.44.2
sentence-transformers==3.1.1

# Vector Database
faiss-cpu==1.8.0.post1

# HTTP Clients
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
openai==1.3.5

# UI Framework
gradio>=4.44.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0

# File Operations
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
